==> notebooks/14_gtex/01_load_positives.ipynb <==
# Load and process fine-mapped GTeX eQTLs (PIP > 0.9) from the eQTL catalog

import numpy as np
import pandas as pd
import anndata
import os
import tqdm

from grelu.data.preprocess import filter_blacklist, filter_chromosomes
from grelu.variant import filter_variants

import torch
%matplotlib inline

## Paths

eqtl_cat_path = 'https://raw.githubusercontent.com/eQTL-Catalogue/eQTL-Catalogue-resources/master/data_tables/dataset_metadata.tsv'
anndata_file = "/gstore/data/resbioai/grelu/decima/20240823/data.h5ad"
susie_dir='/gstore/data/resbioai/grelu/decima/QTS000015/susie/'
out_dir = '/gstore/data/resbioai/grelu/decima/20240823/bulk_eqtl_results/'

## Get GTeX dataset ID to tissue mapping

eqtl_meta = pd.read_table(eqtl_cat_path)
eqtl_meta = eqtl_meta[eqtl_meta.quant_method == 'ge'] # gene exp. QTLs
eqtl_meta = eqtl_meta[eqtl_meta.study_label == 'GTEx']
eqtl_meta = eqtl_meta[['dataset_id', 'sample_group']]

eqtl_meta

ct_dict = eqtl_meta.set_index('dataset_id').sample_group.to_dict()
ct_dict

## Load decima metadata

ad = anndata.read_h5ad(anndata_file)
ensembl_id_map = ad.var[['gene_id']].reset_index().set_index('gene_id')['index'].to_dict()

## Load fine-mapping results

susie_df = []

for ds in ct_dict.keys():
    print(ds)
    df = pd.read_table(f'{susie_dir}/{ds}/{ds}.credible_sets.tsv.gz')
    df['chrom'] = [x.split('_')[0] for x in df.variant]
    df['pos'] = [int(x.split('_')[1]) for x in df.variant]
    df['ref'] = [x.split('_')[2] for x in df.variant]
    df['alt'] = [x.split('_')[3] for x in df.variant]
    df['gene'] = df.gene_id.map(ensembl_id_map)
    df['dataset'] = ds
    df['celltype'] = ct_dict[ds]
    susie_df.append(df)

susie_df = pd.concat(susie_df, axis=0).reset_index(drop=True)
print(len(susie_df))

# make complete list of "credible variants"
cs_vars = set(susie_df['variant'])

print(len(cs_vars))

susie_df.head(3)

## Select positives (PIP > 0.9)

print(len(susie_df))
susie_df = susie_df[susie_df.pip > 0.9]
len(susie_df)

susie_df.dataset.value_counts().sort_values().tail(10)

## Filter fine-mapped variants

susie_df = susie_df[susie_df.gene.notna()]
susie_df = filter_variants(susie_df, max_del_len=0, max_insert_len=0, standard_bases=True) # remove indels
susie_df = filter_chromosomes(susie_df, include='autosomesXY') # keep standard chroms
susie_df = filter_blacklist(susie_df, genome="hg38", window=100) # remove variants in blacklisted regions

susie_df.dataset.value_counts().sort_values().tail(10)

# subset to those in decima intervals
susie_df = susie_df.merge(ad.var[['gene_id', 'start', 'end', 'strand', 'gene_mask_start']]).rename(
    columns={'start': 'gene_window_start', 'end': 'gene_window_end', 'strand': 'gene_strand'}) # add window information
susie_df = susie_df[((susie_df.pos > susie_df.gene_window_start) & (susie_df.pos < susie_df.gene_window_end))] # keep variants within the sequence window
len(susie_df)

susie_df.dataset.value_counts().sort_values().tail(10)

## Add relative position to fine-mapped variants

susie_df['pos_relative'] = susie_df.pos - susie_df.gene_window_start - 1
susie_df.loc[susie_df.gene_strand=='-', 'pos_relative'] = susie_df.gene_window_end[susie_df.gene_strand=='-'] - susie_df.pos[susie_df.gene_strand=='-']
susie_df['abspos_rel_TSS'] = np.abs(susie_df["pos_relative"] - susie_df['gene_mask_start'])

susie_df.head()

## Save

susie_df.to_csv(os.path.join(out_dir, 'susie_df.csv'), index=False)

np.save(os.path.join(out_dir, 'cs_vars.npy'), list(cs_vars))
