==> notebooks/14_gtex/02_load_negatives.ipynb <==
## Load and process negative GTeX SNPs from the eQTL catalog

import numpy as np
import pandas as pd
import polars as pl
import anndata
import os
import tqdm

%matplotlib inline

## Paths

eqtl_cat_path = 'https://raw.githubusercontent.com/eQTL-Catalogue/eQTL-Catalogue-resources/master/data_tables/dataset_metadata.tsv'
anndata_file = "/gstore/data/resbioai/grelu/decima/20240823/data.h5ad"
susie_file='/gstore/data/resbioai/grelu/decima/20240823/bulk_eqtl_results/susie_df.csv'
cs_vars_file='/gstore/data/resbioai/grelu/decima/20240823/bulk_eqtl_results/cs_vars.npy'
out_dir = '/gstore/data/resbioai/grelu/decima/20240823/bulk_eqtl_results/'
sumstats_dir ='/gstore/data/resbioai/grelu/decima/QTS000015/sumstats/'

## Get GTeX dataset ID to tissue mapping

eqtl_meta = pd.read_table(eqtl_cat_path)
eqtl_meta = eqtl_meta[eqtl_meta.quant_method == 'ge'] # gene exp. QTLs
eqtl_meta = eqtl_meta[eqtl_meta.study_label == 'GTEx']
eqtl_meta = eqtl_meta[['dataset_id', 'sample_group']]

eqtl_meta

ct_dict = eqtl_meta.set_index('dataset_id').sample_group.to_dict()
ct_dict

## Load decima metadata

ad = anndata.read_h5ad(anndata_file)
ensembl_id_map = ad.var[['gene_id']].reset_index().set_index('gene_id')['index'].to_dict()

## Load negative variants

# for each gene, we need the interval, to exclude "unseeable" stuff
gene_to_interval_dict = {v.gene_id:{'chr':v.chrom,'start':v.start,'end':v.end} for k,v in ad.var.iterrows()}

def check_scoreability(row):
    if row['gene_id'] not in gene_to_interval_dict:
        return False
    interval = gene_to_interval_dict[row['gene_id']]
    pos = row['position']
    gene_window_start = interval['start']
    gene_window_end = interval['end']
    return (pos > gene_window_start) & (pos < gene_window_end)

susie_df = pd.read_csv(susie_file)
cs_vars = np.load(cs_vars_file)
len(cs_vars)

negvar_list = []

for ds in ct_dict.keys():
    print(ds)

    print("Loading sumstats")
    file = f'{sumstats_dir}/{ds}/{ds}.all.tsv.gz'
    negvars = pl.read_csv(file, separator="\t", 
                      schema_overrides={'chromosome':pl.String},
                      columns=(1,2,3,4,5,7,8,11,16,18),
                     )
    print(len(negvars))

    print("Filtering negatives")
    # Select common SNPs
    negvars = negvars.filter(
        (pl.col('pvalue') > 0.05) & (pl.col('type') == "SNP") & (pl.col('maf') > 0.05)) # pre-select nonsignificant
    # For the same genes as the positives
    negvars = negvars.filter(pl.col('gene_id').is_in(susie_df[susie_df.dataset == ds].gene_id))
    # Not in any credible set
    negvars = negvars.filter(~pl.col('variant').is_in(list(cs_vars)))
    # In decima interval
    negvars = negvars.filter(pl.struct(pl.all()).map_elements(check_scoreability, return_dtype=pl.Boolean))
    
    negvars = negvars.to_pandas()

    print("Filtering chromosomes")
    negvars = negvars.rename(columns={'chromosome':'chrom', 'position':'pos'})
    negvars = negvars.dropna(subset=['chrom'])
    negvars['chrom'] = ['chr' + x for x in negvars.chrom]

    print("Adding metadata")
    negvars['dataset'] = ds
    negvars['celltype'] = ct_dict[ds]

    print(len(negvars))
    negvar_list.append(negvars)

negvars = pd.concat(negvar_list)

## Add gene metadata

# for each negative, compute the relpos and distance to the TSS 
negvar_dedup = negvars[['gene_id','variant','pos']].drop_duplicates()

# add gene information and calculate relative variant positions (offset)
negvar_dedup = negvar_dedup.merge(ad.var[['gene_id', 'start', 'end', 'strand', 'gene_mask_start']]).rename(
    columns={'start': 'gene_window_start', 'end': 'gene_window_end', 'strand': 'gene_strand'}) # add window information
negvar_dedup['pos_relative'] = negvar_dedup.pos - negvar_dedup.gene_window_start - 1

# use gene_end to calculate offset for - genes 
negvar_dedup.loc[negvar_dedup.gene_strand=='-', 'pos_relative'] = negvar_dedup.gene_window_end[negvar_dedup.gene_strand=='-'] - negvar_dedup.pos[negvar_dedup.gene_strand=='-']
negvar_dedup['abspos_rel_TSS'] = np.abs(negvar_dedup["pos_relative"] - negvar_dedup['gene_mask_start'])

negvars = negvars.merge(negvar_dedup[['gene_id','variant','gene_strand',"pos_relative","abspos_rel_TSS"]],
    on=['gene_id','variant'])

## Add gene name from Decima metadata

negvars = negvars.drop(columns=['type'])
negvars['gene'] = negvars.gene_id.map(ensembl_id_map)
negvars.head()

## Add credible set columns

negvars['pip'] = 0
negvars['cs_id'] = 'negative'
negvars['cs_size'] = 0
negvars.head()

## Match to positive variants by distance

# for each positive variant, collect target_negative_n negatives wich are as close to the TSS as possible, and not yet selected for this cell-type
target_negative_n = 20
selected_vars = {}
for celltype in set(susie_df['celltype']):
    print(celltype)
    
    curr_positive_df = susie_df.query('celltype == @celltype')
    curr_negative_df = negvars.query('celltype == @celltype')
    
    selected_vars[celltype] = set()
    for _,positive in curr_positive_df.iterrows():
        negative_sub = curr_negative_df[curr_negative_df.gene_id == positive['gene_id']]
        negative_sub = negative_sub.sort_values('abspos_rel_TSS')
        i = 0
        for _,variant in negative_sub.iterrows():
            if variant['variant'] not in selected_vars[celltype]:
                selected_vars[celltype].add(variant['variant'])
                i += 1
                if i == target_negative_n:
                    break

negvars_reduced = pd.concat([
    negvars[negvars.variant.isin(selected_vars[celltype]) & (negvars.celltype == celltype)
        ] for celltype in selected_vars])
len(susie_df), len(negvars_reduced)

## Print summary stats

rows = []
for celltype in selected_vars:
    rows.append({'celltype':celltype, 
                 'pos_genes':len(set(susie_df.query('celltype == @celltype')['gene'])),
                 'pos_genes_with_neg_ct':len(set(negvars.query('celltype == @celltype')['gene']) & set(susie_df.query('celltype == @celltype')['gene'])),
                 'total_pos':len(set(susie_df.query('celltype == @celltype')['variant'])),
                 'total_matched_neg':len(set(negvars.query('celltype == @celltype').merge(susie_df.query('celltype == @celltype')['gene_id'],on='gene_id')['variant'])),
                 'reduced_matched_neg':len(selected_vars[celltype]),
                 })
    
pd.DataFrame(rows)

## Combine positive and negative variants

susie_df = pd.concat([susie_df, negvars_reduced])
len(susie_df)

susie_df.drop(columns=['molecular_trait_id', 'maf'], inplace=True)

susie_df.head()

## Create VEP input file

variant_df = susie_df[['chrom', 'pos', 'ref', 'alt', 'variant', 'rsid', 'gene_id', 'gene', 'gene_strand', 'pos_relative']].drop_duplicates(subset=['gene_id','variant'])
len(variant_df)

variant_df.head()

## Save final variants

variant_df.to_csv("variants_df.csv", index=False)
susie_df.to_csv("susie_df.csv", index=False)

len(variant_df)
