==> notebooks/14_gtex/05_aggregate_decima.ipynb <==
# Aggregate Decima VEP results

import pandas as pd
import numpy as np
import scanpy as sc
import anndata
import os

## Paths

decima_metadata_file = '/gstore/data/resbioai/grelu/decima/20240823/data.h5ad'
out_dir = '/gstore/data/resbioai/grelu/decima/20240823/bulk_eqtl_results/'

## Load Decima metadata

decima_tasks = sc.read(decima_metadata_file).obs

## Load individual replicate predictions

preds = []
for rep in range(4):
    rep_preds_file = os.path.join(out_dir, f'gtex_eqtl_cat_decima_rep{rep}.pq')
    preds.append(pd.read_parquet(rep_preds_file))

## Split variant metadata from predictions

obs = preds[0].iloc[:, :15]
obs.head()

preds = [p.iloc[:, 15:].values for p in preds]

## Average across replicates

preds = np.stack(preds).mean(0)
preds.shape

## Make anndata

preds = anndata.AnnData(X=preds, obs=obs[['variant', 'gene']], var=decima_tasks)
preds.shape

## Minimally annotate .obs

susie_file = os.path.join(out_dir, "susie_df.csv")
susie_df = pd.read_csv(susie_file)
susie_df=susie_df[['variant', 'pip', 'beta', 'gene', 'celltype']]
susie_df['label'] = susie_df.pip > 0.9
print(len(susie_df))

pos = susie_df[susie_df.label][['variant', 'gene']].drop_duplicates()
pos["is_eQTL"] = True
len(pos)

preds.obs = preds.obs.merge(pos, how="left")
preds.obs.loc[preds.obs.is_eQTL.isna(), "is_eQTL"] = False

preds.obs.is_eQTL = preds.obs.is_eQTL.astype(str)

## Save

out_file = os.path.join(out_dir, 'gtex_eqtl_cat_decima.h5ad')
preds.write_h5ad(out_file)
