==> notebooks/14_gtex/06_aggregate_borzoi.ipynb <==
# Aggregate Borzoi VEP results

import pandas as pd
import numpy as np
import scanpy as sc
import anndata
import os
from tqdm import tqdm

## Paths

preds_dir = '/gstore/data/resbioai/grelu/decima/20240823/bulk_eqtl_results'
borzoi_tasks_file= 'https://github.com/calico/borzoi/raw/refs/heads/main/data/targets_human.txt.gz'
variants_file = os.path.join(preds_dir, "variants_df.csv")

## Combine chunked borzoi results

[x for x in os.listdir(preds_dir) if (x.startswith('gtex_eqtl_cat_borzoi')) & (x.endswith('npy'))]

starts = list(range(0, 229828, 20000))
ends = list(range(20000, 229828, 20000)) + [229828]
chunks = list(zip(starts, ends))
print(chunks)

preds = []
for chunk in tqdm(chunks):
    chunk_preds = []
    for rep in range(4):
        f = os.path.join(preds_dir, f'gtex_eqtl_cat_borzoi_{rep}_{chunk[0]}_{chunk[1]}.npy')
        chunk_preds.append(np.load(f))
    chunk_preds = np.stack(chunk_preds).mean(0)
    preds.append(chunk_preds)

preds = np.vstack(preds)
preds.shape

## Add task and variant metadata

borzoi_tasks = pd.read_table(borzoi_tasks_file, index_col=0)

variants = pd.read_csv(variants_file)
print(len(variants))
variants.head()

preds = anndata.AnnData(X = preds, obs=variants[['variant', 'gene']], var=borzoi_tasks)

## Save

out_file = os.path.join(preds_dir, 'gtex_eqtl_cat_borzoi.h5ad')
preds.write_h5ad(out_file)
