==> notebooks/14_gtex/07_analyze.ipynb <==
# Generate VEP scores for GTeX tissues

import pandas as pd
import numpy as np
import scanpy as sc
import anndata
import os

## Paths

out_dir = '/gstore/data/resbioai/grelu/decima/20240823/bulk_eqtl_results'
susie_file = os.path.join(out_dir, "susie_df.csv")
decima_preds_file = os.path.join(out_dir, 'gtex_eqtl_cat_decima.h5ad')
borzoi_preds_file = os.path.join(out_dir, 'gtex_eqtl_cat_borzoi.h5ad')
gtex_sample_file = 'https://raw.githubusercontent.com/broadinstitute/gtex-v8/master/data/GTEx_Analysis_v8_RNAseq_samples.txt'

## Load fine-mapping results

susie_df = pd.read_csv(susie_file)
susie_df=susie_df[['variant', 'pip', 'beta', 'gene', 'celltype', 'abspos_rel_TSS']]
print(len(susie_df))

## Label positive variants (PIP > 0.9)

susie_df['label'] = susie_df.pip > 0.9
susie_df = susie_df.drop(columns='pip')
susie_df.head(3)

## Load predictions

decima_preds = sc.read(decima_preds_file)
# Only consider healthy tracks to match Gtex
decima_preds = decima_preds[:, decima_preds.var.disease.isin(['healthy', 'NA'])]
decima_preds

borzoi_preds = sc.read(borzoi_preds_file)
borzoi_preds

## Map GTeX tissues to decima tasks

gtex_map = {
 'adipose_subcutaneous': ('tissue', ['adipose tissue', 'subcutaneous adipose tissue']),
 'adrenal_gland': ('tissue', ['adrenal gland']),
 'artery_aorta': ('tissue', ['aorta']),
 'blood':('tissue', ['blood']),
 'brain_amygdala': ('subregion', ['Amygdala']),
 'brain_anterior_cingulate_cortex': ('subregion', ['Anterior cingulate cortex','Anterior cingulate cortex - ACC']),
 'brain_cerebellar_hemisphere': ('tissue', ['Cerebellum_Lateral hemisphere of cerebellum - CBL']),
 'brain_cerebellum': ('region', ['Cerebellum']),
 'brain_cortex': ('region', ['Cerebral cortex']),
 'brain_frontal_cortex': ('tissue',['Cerebral cortex_Frontal Cortex']),
 'brain_hippocampus': ('region',['Hippocampus']),
 'brain_hypothalamus': ('region', ['Hypothalamus']),
 'brain_nucleus_accumbens': ('subregion', ['Nucleus accumbens', 'Nucleus Accumbens - NAC']),
 'brain_putamen':('subregion', ['Putamen - Pu']),
 'brain_spinal_cord': ('region',['Spinal cord']),
 'brain_substantia_nigra': ('subregion', ['Substantia Nigra - SN','Substantia nigra']),
 'breast': ('tissue', ['breast', 'mammary gland']),
 'colon_sigmoid': ('tissue',['colon', 'sigmoid colon']),
 'colon_transverse': ('tissue',['colon', 'transverse colon']),
 'esophagus_muscularis': ('tissue',['esophagus muscularis mucosa']),
 'heart_left_ventricle': ('tissue', ['heart left ventricle']),
 'kidney_cortex': ('tissue',['cortex of kidney']),
 'liver': ('organ', ['liver']),
 'lung': ('organ', ['lung']),
 'muscle': ('tissue', ['muscle tissue', 'psoas muscle']),
 'ovary': ('tissue', ['ovary']),
 'pancreas': ('tissue', ['pancreas']),
 'prostate': ('tissue', ['prostate gland', 'peripheral zone of prostate', 'transition zone of prostate']),
 'small_intestine': ('tissue',['ileum', 'small intestine']),
 'spleen': ('organ', ['spleen']),
 'stomach': ('tissue', ['stomach']),
 'testis': ('organ', ['testis']),
 'uterus': ('organ', ['uterus']),
}

#'adipose_visceral', 'brain_caudate','esophagus_mucosa',
#'minor_salivary_gland', 'pituitary', 'heart_atrial_appendage', 
#'skin_not_sun_exposed', 'skin_sun_exposed', 'thyroid', 'vagina'

gtex_map = pd.DataFrame(gtex_map, index=['decima_col', 'decima_sel']).T.reset_index(names='tissue')
print(len(gtex_map))
gtex_map.head()

## Match these GTeX tissues to Borzoi tracks

gtex_tissues = pd.read_table(gtex_sample_file).set_index('sample_id').tissue_id.to_dict()
gtex_tissues

borzoi_gtex_samples = borzoi_preds.var[borzoi_preds.var.identifier.str.contains('GTEX')].identifier
borzoi_gtex_samples = borzoi_gtex_samples.str.split('.').apply(lambda x: x[0])
borzoi_gtex_samples.head()

borzoi_map = borzoi_gtex_samples.map(gtex_tissues).reset_index().dropna().groupby('identifier').index.apply(list)

borzoi_map.index = [x.lower() for x in borzoi_map.index]

borzoi_map.index.difference(gtex_map.tissue)

set(gtex_map.tissue).difference(borzoi_map.index)

gtex_tissue_to_eqtl_cat_mapping = {
    'adipose_visceral_omentum':'adipose_visceral',
    'whole_blood':'blood',
    'small_intestine_terminal_ileum':'small_intestine',
    'breast_mammary_tissue':'breast',
    'muscle_skeletal':'muscle',
}

for k, v in gtex_tissue_to_eqtl_cat_mapping.items():
    borzoi_map.index = [x if x!=k else v for x in borzoi_map.index]

borzoi_map = pd.DataFrame(borzoi_map).reset_index()
borzoi_map.columns=['tissue', 'borzoi_tracks']

borzoi_map

## Combine decima and borzoi mappings

gtex_map = gtex_map.merge(borzoi_map, how='left')
gtex_map.head(3)

len(gtex_map), len(gtex_map.dropna(subset='borzoi_tracks'))

## Compute scores

def compute_mean_score(ad):
    return ad.X.mean(1)

def compute_weighted_score(ad):
    return np.matmul(ad.X, ad.var.n_cells)/ad.var.n_cells.sum()

decima_scores = dict()
borzoi_scores = dict()

# For each tissue
for row in gtex_map.itertuples():

    # Compute decima score
    decima_sub = decima_preds[:, decima_preds.var[row.decima_col].isin(row.decima_sel)]
    decima_scores[row.tissue] = compute_weighted_score(decima_sub)
    # Compute Borzoi score
    if isinstance(row.borzoi_tracks, list):
        borzoi_scores[row.tissue] =  compute_mean_score(borzoi_preds[:, row.borzoi_tracks])
    
decima_scores = pd.DataFrame(decima_scores)
borzoi_scores = pd.DataFrame(borzoi_scores)

decima_scores = pd.concat([decima_preds.obs.reset_index(drop=True), decima_scores], axis=1)
borzoi_scores = pd.concat([borzoi_preds.obs.reset_index(drop=True), borzoi_scores], axis=1)

decima_scores.head(2)

## Save scores

decima_scores.to_csv(
    os.path.join(out_dir, 'gtex_eqtl_cat_decima_scores.csv'),
    index=False
)
borzoi_scores.to_csv(
    os.path.join(out_dir, 'gtex_eqtl_cat_borzoi_scores.csv'),
    index=False
)

## Subset susie to represented tissues

print(len(susie_df))
susie_df = susie_df[susie_df.celltype.isin(gtex_map.tissue)]
print(len(susie_df))

## Merge scores with fine-mapping results

decima_scores = decima_scores.melt(
    id_vars=['variant', 'gene'], var_name='celltype', value_name='decima_score'
)
borzoi_scores = borzoi_scores.melt(
    id_vars=['variant', 'gene'], var_name='celltype', value_name='borzoi_score'
)

decima_scores.head(3)

print(len(susie_df))
susie_df = susie_df.merge(decima_scores, on=['variant', 'gene', 'celltype'], how='left')
susie_df = susie_df.merge(borzoi_scores, on=['variant', 'gene', 'celltype'], how='left')
print(len(susie_df))

## Compute absolute scores

susie_df['abs_decima_score'] = np.abs(susie_df['decima_score'])
susie_df['abs_borzoi_score'] = np.abs(susie_df['borzoi_score'])

## Save

susie_df.to_csv(os.path.join(out_dir, 'gtex_eqtl_cat_susie_scored.csv'),
               index=False)
